Recaptcha: -updated downloadCaptcha to return null because download will fail with 400 Bad Request ChallengeResponseController: -since RC1 shutdown only a dummy response with two words is required PluginForHost/PluginForDecrypt: -updated copyCaptcha to handle null Plugin: -updated clean to handle null TwentyOneMembersCom: -fixed login -fixes ticket GGE-671-15676 build>full
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@39062 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/plugins/Plugin.java
@@ -515,7 +515,7 @@ public abstract class Plugin implements ActionListener {
         cleanupLastChallengeResponse();
         br = null;
         for (final File clean : cleanUpCaptchaFiles) {
-            if (!clean.delete()) {
+            if (clean != null && !clean.delete()) {
                 clean.deleteOnExit();
             }
         }
